move linux to sdl_wrapper - as it isn't specific to linux, and works in windows as well
